github-alert-analyzer/backend/requirements.txt
# FastAPI and ASGI
fastapi==0.124.0
starlette==0.49.1
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Authentication & Security
pyjwt[crypto]==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
cryptography==44.0.1

# HTTP & API clients
httpx==0.28.1
aiohttp==3.12.14

# Validation & Settings
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0


# LLM Providers
anthropic==0.40.0
openai==1.57.2

# Utilities
python-dotenv==1.0.1
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Security scanning
pip-audit==2.7.3
bandit==1.8.0
